refactor: update html , css, js + some be code =))

+++ backend/src/main/java/com/darian/ecommerce/businesslogic/shippingfee/RushShippingFeeCalculator.java
@@ -5,8 +5,8 @@ import com.darian.ecommerce.dto.BaseOrderDTO;
 public class RushShippingFeeCalculator implements ShippingFeeCalculator{
     // Calculate rush shipping fee (example logic)
     @Override
-    public Integer calculateShippingFee(BaseOrderDTO dto) {
+    public Float calculateShippingFee(BaseOrderDTO dto) {
         // Higher base fee for rush delivery
-        return 20 + (dto.getSubtotal() > 100 ? 10 : 0);
+        return 20f + (dto.getSubtotal() > 100 ? 10f : 0f);
     }
 }
