+++ backend/src/main/java/com/darian/ecommerce/businesslogic/shippingfee/StandardShippingFeeCalculator.java
@@ -5,8 +5,8 @@ import com.darian.ecommerce.dto.BaseOrderDTO;
 public class StandardShippingFeeCalculator  implements ShippingFeeCalculator{
     // Calculate standard shipping fee (example logic)
     @Override
-    public Integer calculateShippingFee(BaseOrderDTO dto) {
+    public Float calculateShippingFee(BaseOrderDTO dto) {
         // Base fee + additional based on subtotal
-        return 10 + (dto.getSubtotal() > 100 ? 5 : 0);
+        return 10f + (dto.getSubtotal() > 100 ? 5f : 0f);
     }
 }
